Add polygon mappingg for booking

+++ Tawuniya.Web/Tawuniya.Web/Tawuniya.Web/Models/Seats/SeatViewModel.cs
@@ -3,7 +3,7 @@
     public class SeatViewModel
     {
         public int LayoutId { get; set; }
-        public float X { get; set; }
-        public float Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
     }
 }
